Refact : 1주차 금요일 리뷰 반영
