+++ FE/issuetracker/src/components/Issue/IssueTitleButton.tsx
@@ -1,0 +1,16 @@
+import * as S from 'components/Issue/styled.titleButton';
+
+interface Props {
+  buttonIcon: React.ReactNode;
+  buttonText: string;
+}
+
+function IssueTitleButton({ buttonIcon, buttonText }: Props) {
+  return (
+    <S.issueButton>
+      {buttonIcon}
+      <S.buttonText>{buttonText}</S.buttonText>
+    </S.issueButton>
+  );
+}
+export default IssueTitleButton;
